Enable hosted client by default
