+++ src/Orleans.Runtime/Core/HostedClient.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Threading;
@@ -13,7 +13,7 @@ namespace Orleans.Runtime
     /// <summary>
     /// A client which is hosted within a silo.
     /// </summary>
-    internal sealed class HostedClient : IDisposable, IHostedClient
+    internal sealed class HostedClient : IDisposable, IHostedClient, ILifecycleParticipant<ISiloLifecycle>
     {
         private readonly BlockingCollection<Message> incomingMessages = new BlockingCollection<Message>();
         private readonly CancellationTokenSource listeningCts = new CancellationTokenSource();
@@ -48,14 +48,6 @@ namespace Orleans.Runtime
             this.logger = logger;
 
             this.ClientAddress = ActivationAddress.NewActivationAddress(siloDetails.SiloAddress, GrainId.NewClientId());
-
-            // Register with the directory and message center so that we can receive messages.
-            this.clientObserverRegistrar.SetHostedClient(this);
-            this.clientObserverRegistrar.ClientAdded(this.ClientId);
-            this.siloMessageCenter.SetHostedClient(this);
-
-            // Start pumping messages.
-            this.Start();
         }
 
         /// <inheritdoc />
@@ -201,6 +193,28 @@ namespace Orleans.Runtime
                 {
                     this.logger.Error(ErrorCode.Runtime_Error_100326, "RunClientMessagePump has thrown exception", exc);
                 }
+            }
+        }
+
+        void ILifecycleParticipant<ISiloLifecycle>.Participate(ISiloLifecycle lifecycle)
+        {
+            lifecycle.Subscribe("HostedClient", ServiceLifecycleStage.BecomeActive, OnStart, OnStop);
+
+            Task OnStart(CancellationToken cancellation)
+            {
+                // Register with the directory and message center so that we can receive messages.
+                this.clientObserverRegistrar.SetHostedClient(this);
+                this.clientObserverRegistrar.ClientAdded(this.ClientId);
+                this.siloMessageCenter.SetHostedClient(this);
+
+                // Start pumping messages.
+                this.Start();
+                return Task.CompletedTask;
+            }
+
+            Task OnStop(CancellationToken cancellation)
+            {
+                return Task.CompletedTask;
             }
         }
     }
