+++ src/Orleans.Runtime/Hosting/DefaultSiloServices.cs
@@ -274,8 +274,21 @@ namespace Orleans.Hosting
             services.AddTransient<IConfigurationValidator, ClusterOptionsValidator>();
             services.AddTransient<IConfigurationValidator, SiloClusteringValidator>();
 
-            // Disable hosted client by default.
-            services.TryAddSingleton<IHostedClient, DisabledHostedClient>();
+            // Enable hosted client.
+            services.AddSingleton<HostedClient>();
+            services.AddFromExisting<IHostedClient, HostedClient>();
+            services.AddFromExisting<ILifecycleParticipant<ISiloLifecycle>, HostedClient>();
+            services.TryAddSingleton<InvokableObjectManager>();
+            services.TryAddSingleton<IClusterClient>(
+                sp =>
+                {
+                    var result = ActivatorUtilities.CreateInstance<ClusterClient>(sp);
+                    var task = result.Connect();
+                    if (!task.IsCompleted)
+                        throw new InvalidOperationException(
+                            $"{nameof(result.Connect)}() method for internal {nameof(IClusterClient)} must complete synchronously.");
+                    return result;
+                });
 
             // Enable collection specific Age limits
             services.AddOptions<GrainCollectionOptions>()
